+++ ChkdraftLib/Windows/DialogWindows/MapSettings/Forces.cpp
@@ -38,7 +38,7 @@ ForcesWindow::~ForcesWindow()
 
 }
 
-bool ForcesWindow::CreateThis(HWND hParent, u32 windowId)
+bool ForcesWindow::CreateThis(HWND hParent, u64 windowId)
 {
     if ( getHandle() != NULL )
         return SetParent(hParent);
