+++ ChkdraftLib/Windows/DialogWindows/MapSettings/UpgradeSettings.cpp
@@ -52,7 +52,7 @@ UpgradeSettingsWindow::~UpgradeSettingsWindow()
 
 }
 
-bool UpgradeSettingsWindow::CreateThis(HWND hParent, u32 windowId)
+bool UpgradeSettingsWindow::CreateThis(HWND hParent, u64 windowId)
 {
     if ( getHandle() != NULL )
         return SetParent(hParent);
