+++ ChkdraftLib/Windows/DialogWindows/TrigModify/TrigConditions.cpp
@@ -26,7 +26,7 @@ TrigConditionsWindow::~TrigConditionsWindow()
 
 }
 
-bool TrigConditionsWindow::CreateThis(HWND hParent, u32 windowId)
+bool TrigConditionsWindow::CreateThis(HWND hParent, u64 windowId)
 {
     if ( getHandle() != NULL )
         return SetParent(hParent);
