+++ ChkdraftLib/Windows/DialogWindows/TrigModify/TrigGeneral.cpp
@@ -38,7 +38,7 @@ TrigGeneralWindow::~TrigGeneralWindow()
 
 }
 
-bool TrigGeneralWindow::CreateThis(HWND hParent, u32 windowId)
+bool TrigGeneralWindow::CreateThis(HWND hParent, u64 windowId)
 {
     if ( getHandle() != NULL )
         return SetParent(hParent);
