+++ WindowsLib/DragListControl.h
@@ -8,7 +8,7 @@ namespace WinLib {
     {
         public:
             virtual ~DragListControl();
-            bool CreateThis(HWND hParent, int x, int y, int width, int height, u32 id);
+            bool CreateThis(HWND hParent, int x, int y, int width, int height, u64 id);
     };
 
 }
